# src/swagger.json
{
	"components": {
...
		"schemas": {
...
			"UserRole": {
				"type": "string",
				"enum": [
					"TST",
					"MP",
					"ADMIN"
				],
				"description": "Enums"
			},
			"UserInfo": {
				"properties": {
					"id": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					}
				},
				"required": [
					"id",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"token": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "loanhood-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/auth/my-profile": {
			"get": {
				"operationId": "GetMyProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/UserInfo"
										},
										{
											"properties": {
												"email": {
													"type": "string"
												},
												"role": {
													"$ref": "#/components/schemas/UserRole"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"email",
												"role",
												"id"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"token": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}